Game bug fixed
