+++ Utils.py
@@ -38,15 +38,4 @@ def create_ship_type_list(args):
     return type_list
 
 
-def is_win(ship_list):
-    """
-    :param ship_list:the list of ships of the enemy player
-    :return: a logical value that returns true if the player won the game
-    """
-    j = 0
-    is_alive = False
-    while j < len(ship_list) and not is_alive:
-        if not (ship_types.Ship.is_sunk(ship_list[j])):
-            is_alive = True
-        j = j + 1
-    return not is_alive
+
