+++ main.py
@@ -5,21 +5,18 @@ import inputs
 import game
 import Utils
 
+
 args = inputs.initialize_parser()
 
 type_list = Utils.create_ship_type_list(args)
 board_player1, ship_list1 = game_board.create_board(args, type_list)
 print('\n\n\n\nPass now the computer to the other player!')
 board_player2, ship_list2 = game_board.create_board(args, type_list)
-ship_list=[]
-ship_list.append(ship_list1)
-ship_list.append(ship_list2)
-context = game.Context(game.Player1())#Player1 starts the game
+#Start the game, player_1 starts
 for i in range(args.turns):
-    context.choose_and_shoot(ship_list)
+    game.player1_shoot(ship_list2)
+    game.player2_shoot(ship_list1)
 
-#Todo: gameplay to be fixed
-#game.gameplay(args, ship_list)
